- change fp8 to int8 - load llama config from local files

+++ dps/utils/precision.py
@@ -34,7 +34,7 @@ def map_to_dtype(precision: Precision):
     elif precision == Precision.BFLOAT16:
         dtype = torch.bfloat16
     elif precision == Precision.FP8:
-        dtype = torch.float8
+        dtype = torch.int8
     elif precision == Precision.INT4:
         dtype = torch.int4
     elif precision == Precision.INT2:
